ceph external bluestore db devices detection #12463
This changes how find_disks explores ceph drives. It will handle cases
of bluestore when using between 1 and 3 devices for various roles.
external_journal and external_bluedb means that there exists a dedicated
partition which should be used only for that role. So the meaning of
'external' has changed now. This might have breaken ceph setup without
bluestore, but we don't need filestore.